cleaner/requirements.txt
redis==4.1.1
mysql-connector-python==8.0.28

